+++ fist-kit-infra/fist-trace/fist-trace-context/src/main/java/com/power4j/fist/trace/context/item/TraceContextItemFactory.java
@@ -1,0 +1,15 @@
+package com.power4j.fist.trace.context.item;
+
+/**
+ * 追踪上下文项工厂。
+ *
+ * @author CJ (power4j@outlook.com)
+ * @since 3.15
+ */
+public interface TraceContextItemFactory {
+
+	String processor();
+
+	TraceContextItem create(TraceContextItemSpec spec, TraceContextItemFactoryContext context);
+
+}